Wrap Projects page with new ContentContainerNoSidebar component
