+++ client/src/components/Projects/ProjectsPage.js
@@ -13,21 +13,9 @@ import DeleteIcon from "../../images/trash.png";
 import Pagination from "../Pagination.js";
 import DeleteProjectModal from "./DeleteProjectModal";
 import DuplicateProjectModal from "./DuplicateProjectModal";
-import {
-  useAppInsightsContext,
-  useTrackMetric
-} from "@microsoft/applicationinsights-react-js";
+import ContentContainerNoSidebar from "../Layout/ContentContainerNoSidebar";
 
 const useStyles = createUseStyles({
-  main: {
-    display: "flex",
-    flexDirection: "column",
-    justifyContent: "flex-start",
-    alignItems: "center",
-    minHeight: "calc(100vh - 103px - 48px)",
-    margin: "auto",
-    width: "85%"
-  },
   pageTitle: {
     marginTop: "2em"
   },
@@ -131,11 +119,6 @@ const ProjectsPage = ({ account, history }) => {
   const currentUser = JSON.parse(localStorage.getItem("currentUser"));
   const toastAdd = toast.add;
   const historyPush = history.push;
-
-  const appInsights = useAppInsightsContext();
-
-  //appInsights.trackMetric("ProjectPage Component");
-  const trackComponent = useTrackMetric(appInsights, "ProjectsPage");
 
   const pageLinks = document.getElementsByClassName("pageLinkContainer-0-2-40");
   for (let i = 0; i < pageLinks.length; i++) {
@@ -312,11 +295,7 @@ const ProjectsPage = ({ account, history }) => {
   );
 
   return (
-    <div
-      className={classes.main}
-      onLoad={trackComponent}
-      onClick={trackComponent}
-    >
+    <ContentContainerNoSidebar componentToTrack="ProjectsPage">
       <h1 className={classes.pageTitle}>Projects</h1>
       <div className={classes.searchBarWrapper}>
         <input
@@ -470,7 +449,7 @@ const ProjectsPage = ({ account, history }) => {
           />
         </>
       )}
-    </div>
+    </ContentContainerNoSidebar>
   );
 };
